+++ app/client/app/app-entry.ts
@@ -6,7 +6,7 @@ import { Router } from '@lit-labs/router';
 import { reset } from '@/styles/reset.styles';
 
 import '@/app/app-frame';
-import { routes } from '@/routes';
+import { routes } from '@/router/routes';
 
 @customElement('app-entry')
 export class AppEntry extends LitElement {
